Fixes #107
The information about if a sound was annotated completely was store previously int he sound object.
As now we want to show this information accroding to the user that is logged in a Complete model is created to inform if a user has completed a sound.
A method is added in Sound model to check if there are as many SimilarityAnnotation objects as Annotation objects for the sound for the inout user.
In the sounds_list view we create a list of with the sounds and if they are complete for the user.